make branch_revs Git method instead of backend
