+++ dvc/repo/experiments/utils.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Generator, Iterable
+from typing import TYPE_CHECKING, Generator, Iterable, Set
 
 from .base import EXEC_NAMESPACE, EXPS_NAMESPACE, EXPS_STASH, ExpRefInfo
 
@@ -75,13 +75,14 @@ def remote_exp_refs_by_baseline(
 
 def exp_commits(
     scm: "Git", ref_infos: Iterable["ExpRefInfo"] = None
-) -> Generator[str, None, None]:
+) -> Iterable[str]:
     """Iterate over all experiment commits."""
-    shas = set()
+    shas: Set["str"] = set()
     refs = ref_infos if ref_infos else exp_refs(scm)
     for ref_info in refs:
         shas.update(scm.branch_revs(str(ref_info), ref_info.baseline_sha))
-        shas.add(ref_info.baseline_sha)
+        if ref_info.baseline_sha:
+            shas.add(ref_info.baseline_sha)
     yield from shas
