+++ dvc/scm/git/backend/dulwich.py
@@ -251,9 +251,6 @@ class DulwichBackend(BaseGitBackend):  # pylint:disable=abstract-method
     def resolve_commit(self, rev: str) -> "GitCommit":
         raise NotImplementedError
 
-    def branch_revs(self, branch: str, end_rev: Optional[str] = None):
-        raise NotImplementedError
-
     def _get_stash(self, ref: str):
         from dulwich.stash import Stash as DulwichStash
